requirements.txt
absl-py==1.2.0
asttokens==2.0.8
backcall==0.2.0
cachetools==5.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.5
contourpy==1.0.5
cryptography==42.0.2
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
executing==1.1.0
fonttools==4.37.4
glove-python-binary==0.2.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
h5py==3.10.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.5.0
jedi==0.18.1
joblib @ file:///C:/b/abs_e60_bwl1v6/croot/joblib_1666298845728/work
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
networkx==3.1
nltk==3.8.1
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prior==1.0.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==2.2.0
Pygments==2.13.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn @ file:///C:/b/abs_e01rh8f1vi/croot/scikit-learn_1675454931501/work
scipy==1.9.3
seaborn==0.12.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
torch==1.8.0
torchaudio==0.8.0
torchnet==0.0.4
torchvision==0.9.0
tornado==6.4
tqdm==4.64.1
traitlets==5.4.0
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3==1.26.12
visdom==0.2.4
wcwidth==0.2.5
websocket-client==1.7.0
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.16.0
zipp==3.8.1
